+++ apps/e2e/text-editor-storybook-e2e/src/e2e/text-editor.component.cy.ts
@@ -158,5 +158,5 @@ describe('text-editor-storybook', () => {
         });
       });
     });
-  });
+  }, true);
 });