+++ apps/playground/src/app/components/forms/file-attachment/file-attachment.component.ts
@@ -18,7 +18,7 @@ import {
   templateUrl: './file-attachment.component.html',
 })
 export class FileAttachmentComponent {
-  public acceptedTypes = 'image/png,image/jpeg';
+  public acceptedTypes = 'image/png,image/jpeg,application/pdf';
 
   public allItems: (SkyFileItem | SkyFileLink)[];