+++ libs/components/lists/src/lib/modules/repeater/repeater.component.spec.ts
@@ -5,7 +5,6 @@ import {
   flush,
   inject,
   tick,
-  waitForAsync,
 } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { SkyAppTestUtility, expect, expectAsync } from '@skyux-sdk/testing';
@@ -409,32 +408,32 @@ describe('Repeater item component', () => {
     expect(repeaterEls[2].getAttribute('aria-selected')).toBe('true');
   }));
 
-  it('should blur item when item is disabled', waitForAsync(() => {
+  it('should blur item when item is disabled', async () => {
     const fixture = TestBed.createComponent(RepeaterTestComponent);
+
     fixture.detectChanges();
-    fixture
-      .whenStable()
-      .then(() => {
-        fixture.nativeElement
-          .querySelector('sky-repeater-item-content a')
-          .focus();
-        expect(
-          fixture.nativeElement
-            .querySelector('sky-repeater')
-            .matches(':focus-within'),
-        ).toBeTruthy();
-        fixture.componentInstance.disableFirstItem = true;
-        fixture.detectChanges();
-        return fixture.whenStable();
-      })
-      .then(() => {
-        expect(
-          fixture.nativeElement
-            .querySelector('sky-repeater')
-            .matches(':focus-within'),
-        ).toBeFalsy();
-      });
-  }));
+
+    await fixture.whenStable();
+
+    fixture.nativeElement.querySelector('sky-repeater-item-content a').focus();
+
+    expect(
+      fixture.nativeElement
+        .querySelector('sky-repeater')
+        .matches(':focus-within'),
+    ).toBeTruthy();
+
+    fixture.componentInstance.disableFirstItem = true;
+
+    fixture.detectChanges();
+    await fixture.whenStable();
+
+    expect(
+      fixture.nativeElement
+        .querySelector('sky-repeater')
+        .matches(':focus-within'),
+    ).toBeFalsy();
+  });
 
   it('should hide the chevron and disable expand/collapse for items with no content', fakeAsync(() => {
     const fixture = TestBed.createComponent(RepeaterTestComponent);