+++ libs/components/split-view/testing/src/public-api.ts
@@ -2,3 +2,10 @@ export { SkySplitViewFixture } from './legacy/split-view/split-view-fixture';
 export { SkySplitViewFixtureDrawer } from './legacy/split-view/split-view-fixture-drawer';
 export { SkySplitViewFixtureWorkspace } from './legacy/split-view/split-view-fixture-workspace';
 export { SkySplitViewTestingModule } from './legacy/split-view/split-view-testing.module';
+
+export { SkySplitViewDrawerHarness } from './modules/split-view/split-view-drawer-harness';
+export { SkySplitViewHarness } from './modules/split-view/split-view-harness';
+export { SkySplitViewHarnessFilters } from './modules/split-view/split-view-harness-filters';
+export { SkySplitViewWorkspaceContentHarness } from './modules/split-view/split-view-workspace-content-harness';
+export { SkySplitViewWorkspaceFooterHarness } from './modules/split-view/split-view-workspace-footer-harness';
+export { SkySplitViewWorkspaceHarness } from './modules/split-view/split-view-workspace-harness';